# bpo-25246: Improve the performance of deque_remove()