--- a/Misc/NEWS.d/next/Core and Builtins/2018-06-13-00-23-19.bpo-25246.HKOrh3.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2018-06-13-00-23-19.bpo-25246.HKOrh3.rst
@@ -1,0 +1,1 @@
+Improve the performance of `collections.deque.remove`.