Performance using perf:

```
Mean +- std dev: [new] 405 us +- 12 us -> [old] 608 us +- 18 us: 1.50x slower (+50%)
Significant (t=-43.93)
```

<!-- issue-number: bpo-25246 -->
https://bugs.python.org/issue25246
<!-- /issue-number -->

Adapted from Serhiy and Raymond code.