@pablogsal 2018-06-15
I need to play a bit with other implementations so I am going to close this PR until I am happy with the final proposal.